Implementando a primeira consulta gratis

+++ app/Services/API/v1/Management/DestroyQueryService.php
@@ -4,7 +4,10 @@
 namespace App\Services\API\v1\Management;
 
 
+use App\Mail\QueryCanceledClient;
 use App\Repositories\QueryRepository;
+use Carbon\Carbon;
+use Illuminate\Support\Facades\Mail;
 
 class DestroyQueryService
 {
@@ -22,6 +25,19 @@ class DestroyQueryService
 
         if(!$query)
             throw new \Exception("Consulta não localizada.", 500);
+
+        $query_hour = Carbon::create($query->day_hour);
+        $now = Carbon::now();
+
+        if($query_hour->diffInHours($now) < 48)
+            throw new \Exception("Esta consulta não pode mais ser cancelada pois está com menos de 48h para sua realização.", 500);
+
+        $data = [
+            'name' => $query->client->name,
+            'email' => $query->client->email,
+        ];
+
+        Mail::send(new QueryCanceledClient($data));
 
         return $this->query_repository->destroy($query);
     }
